\name{Hasselblad1969}
\docType{data}
\alias{poissmix.dat}

\title{FIX}
\description{FIX.}

\usage{
	data(Hasselblad1969)
}

\format{
	A data.frame with two columns ... FIX.
}

\details{
	FIX
}

\source{
	FIX. 
}

\references{
	\emph{FIX}.
	  \url{FIX}.
}

\examples{
 require("SQUAREM")
 data(Hasselblad1969, package="SQUAREM")

 plot(poissmix.dat) 
}

\keyword{datasets}
